[WIP] とりあえずGradationViewに切り出し
